+++ test/UnitTest.hs
@@ -7,6 +7,11 @@ import qualified Test.Fin
 import qualified Test.FinMap
 import qualified Test.List
 import qualified Test.NatRepr
+import qualified Test.SmallArrayF
+import qualified Test.SmallArrayF.Instances
+import qualified Test.SmallArrayF.Rules
+import qualified Test.SmallArrayF.RulesCoverage
+import qualified Test.SmallArrayF.RulesSemantics
 import qualified Test.Some
 import qualified Test.SymbolRepr
 import qualified Test.TH
@@ -29,6 +34,11 @@ tests = testGroup "ParameterizedUtils" <$> sequence
   , Test.Fin.finTests
   , Test.FinMap.finMapTests
   , Test.NatRepr.natTests
+  , Test.SmallArrayF.smallArrayFTests
+  , pure Test.SmallArrayF.Instances.tests
+  , pure Test.SmallArrayF.Rules.tests
+  , pure Test.SmallArrayF.RulesSemantics.tests
+  , pure Test.SmallArrayF.RulesCoverage.tests
   , Test.Some.someTests
   , Test.SymbolRepr.symbolTests
   , Test.TH.thTests